backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
# 仿真依赖（urban_sim 所需）
pydantic>=2.0
litellm>=1.0
tqdm
shapely
pyproj
pycityproto
